+++ sparkle_log/__about__.py
@@ -12,10 +12,11 @@ __all__ = [
     "__readme__",
     "__repository__",
     "__homepage__",
+    "__documentation__",
 ]
 
 __title__ = "sparkle-log"
-__version__ = "0.2.0"
+__version__ = "0.3.0"
 __description__ = "Write sparkline graphs of CPU and memory usage to your logs."
 __author__ = "Matthew Martin"
 __author_email__ = "matthewdeanmartin@gmail.com"
@@ -25,3 +26,4 @@ __license__ = "MIT"
 __readme__ = "README.md"
 __repository__ = "https://github.com/matthewdeanmartin/sparkle_log"
 __homepage__ = "https://github.com/matthewdeanmartin/sparkle_log"
+__documentation__ = "https://matthewdeanmartin.github.io/sparkle_log/sparkle_log/index.html"