+++ sparkle_log/as_context_manager.py
@@ -9,6 +9,7 @@ from typing import Any, Optional
 
 from sparkle_log.graphs import GLOBAL_LOGGER
 from sparkle_log.scheduler import run_scheduler
+from sparkle_log.ui import GraphStyle
 
 
 class MetricsLoggingContext:
@@ -16,16 +17,17 @@ class MetricsLoggingContext:
     Context manager to log system metrics.
     """
 
-    def __init__(self, metrics=("cpu", "memory"), interval: int = 10) -> None:
+    def __init__(self, metrics=("cpu", "memory"), interval: int = 10, style: GraphStyle = "faces") -> None:
         """Initialize the context manager."""
         if not metrics:
             metrics = ("cpu", "memory")
         else:
             for metric in metrics:
-                if metric not in ("cpu", "memory"):
+                if metric not in ("cpu", "memory", "drive"):
                     raise TypeError("Unexpected metric")
         self.metrics = metrics
         self.interval = interval
+        self.style = style
         self.stop_event: Optional[Event] = None
         self.scheduler_thread: Optional[Thread] = None
 
@@ -33,7 +35,9 @@ class MetricsLoggingContext:
         """Start the context manager, if logging enabled."""
         if GLOBAL_LOGGER.isEnabledFor(logging.INFO):
             self.stop_event = Event()
-            self.scheduler_thread = Thread(target=run_scheduler, args=(self.stop_event, self.metrics, self.interval))
+            self.scheduler_thread = Thread(
+                target=run_scheduler, args=(self.stop_event, self.metrics, self.interval, self.style)
+            )
             self.scheduler_thread.start()
         return self