+++ sparkle_log/as_decorator.py
@@ -8,11 +8,12 @@ from threading import Event, Thread
 
 from sparkle_log.graphs import GLOBAL_LOGGER
 from sparkle_log.scheduler import run_scheduler
+from sparkle_log.ui import GraphStyle
 
 INITIALIZED = False
 
 
-def monitor_metrics_on_call(metrics=("cpu", "memory"), interval=10):
+def monitor_metrics_on_call(metrics=("cpu", "memory"), interval=10, style: GraphStyle = "bar"):
     """
     Decorator to monitor the system metrics while the function is being executed.
     """
@@ -26,7 +27,7 @@ def monitor_metrics_on_call(metrics=("cpu", "memory"), interval=10):
             if not GLOBAL_LOGGER.isEnabledFor(logging.INFO):
                 return func(*args, **kwargs)
             stop_event = Event()
-            scheduler_thread = Thread(target=run_scheduler, args=(stop_event, metrics, interval))
+            scheduler_thread = Thread(target=run_scheduler, args=(stop_event, metrics, interval, style))
             scheduler_thread.start()
 
             try: