+++ sparkle_log/graphs.py
@@ -1,4 +1,7 @@
-# Initialize the logger
+"""
+Global logger to avoid cyclical imports
+"""
+
 import logging
 
 GLOBAL_LOGGER = logging.getLogger(__name__)