+++ test.sh
@@ -1,0 +1,4 @@
+set -e
+sparkle_log --help
+sparkle_log --version
+sparkle_log --interval 1 --duration 3 --style linear