+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2024-04-27
+
+### Added
+
+- Supports graph styles: bar, faces, jagged, linear, vertical, ascii_art, pie_chart
+- Supports new metric, "drive" for free drive space
+
+### Fixed
+- Merged fix to upstream library to fix 3.12 warnings.
+
 ## [0.2.0] - 2024-04-24
 
 ### Added